Remove suggestion to buy local
Buying local may be beneficial for various reasons, but this paragraph seems poorly supported and less useful to the audience:

- It is not clear how buying locally will reduce Covid-19 transmission.
- Supporting the local economy is intuitive, but referencing evidence would make this point stronger anyway.
- It seems there is some debate about the effect of buying local on carbon footprint. For example, see [here](https://ourworldindata.org/food-choice-vs-eating-local).

Maybe rewrite this with more evidence for each claim, if desired. However it also seems to me that this paragraph is less helpful to readers who are trying to learn how to survive Covid-19 and help reduce its impact.