Create README.md
This project was launched at the early part of the pandemic when comprehensive guidance was not forthcoming. At this point in 2021, the very best tool in the arsenal is to get vaccinated as soon as you are eligible and don't neglect your second dose. It is the fastest way for everyone to return to normalcy. If vaccines are not available for your area or your age group, continue to practice safe behaviors.